backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.0
httpx==0.27.0
resend==2.0.0
pydantic==2.9.0
